Using EMF to access database

+++ src/main/java/com/group3/budgetApp/repository/UserRepository.java
@@ -6,5 +6,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Long> {
+public interface UserRepository extends JpaRepository<User,Integer> {
+
+
 }
